requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pytest>=7.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
